chore: move navigation file to its own folder to enable further refactoring
